apidoc: Add type annotations and fix column widths.
The availableKeys call was updated to provide the D-Bus type signature
to all the parameters supported in the control API, so we can use a little
javascript to convert them into user-friendly type annotations.

We also found that the column widths in the parameter table were becomming
a bit of a mess, some add some container-fluid class magic to tidy that up.