apidoc: Use parsed docstrings for params, add method dropdown.
The API now presents parsed docstrings for parameters too, which can
break down complex types and enums, so we can try to imrpove the
documentation on complex types. This gets the information out there
but does a bad job of formatting dicts for the time being.

While we're at it, we also build a dropdown navbar for the methods
in the API, which makes seeking around a bit nicer.