Move parameter value column to an collapseable button.
The formatting of the parameter value was getting badly constrained
for space, especially now that some parameters can have an Args/Returns
table for more details. To try and tidy this up, we can move the
current value into the description column, and use a collapse element
to hide complicated types unless desired.